First reimplementation of travel with websockets
